WIP: prepare for protocol refactoring
