+++ src/lib/constants.ts
@@ -1,4 +1,4 @@
-import type { LifewheelState, LifewheelStep, ProtocolVersion, ReflectionStep, TextStep } from './types'
+import type { LifewheelState, LifewheelStep, ReflectionStep, TextStep } from './types'
 
 export const APP_NAME = 'Life Wheel'
 export const APP_TAGLINE = 'Reflect on Your Life Balance'
@@ -163,15 +163,5 @@ export const MAX_LEVEL = 10
  * This is key in enabling the tweened motion.
  */
 export const INITIAL_LIFEWHEEL_STATE: LifewheelState = [0, 0, 0, 0, 0, 0, 0, 0]
-
-/**
- * If the need arises, we could abstract away implementation details with a common public API surface
- */
-export const PROTOCOL_VERSIONS = {
-    1: {},
-    2: {}
-}
-
-export const CURRENT_PROTOCOL_VERSION: ProtocolVersion = 2
 
 export const REPO_URL = 'https://github.com/Greenheart/lifewheel'
