+++ src/lib/protocols/v1.ts
@@ -1,0 +1,10 @@
+import type { Protocol } from ".";
+
+export default {
+    encodeReflectionEntries(reflections) {
+        
+    },
+    decodeReflectionEntries(data) {
+        
+    }
+} satisfies Protocol
